proje/requirements.txt
Django>=5.1.4
psycopg2-binary>=2.9.10
Pillow>=11.1.0
djangorestframework>=3.15.2
firebase-admin>=6.7.0
python-dotenv>=1.0.0 